optionally log to file not stderr
oddly enough, this is for development not for production.  In
production, logging to standard streams is picked up by docker things
and send to centralised log server.  In Cider, sending too much stuff
to stderr seems to confuse the hell out of it, making it simpler to log
to a file and have it shown by a process running under foreman